+++ cardform/src/main/java/com/mercadolibre/android/cardform/data/repository/FinishInscriptionRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.mercadolibre.android.cardform.data.repository
 
-import com.mercadolibre.android.cardform.base.BaseCoroutine
+import com.mercadolibre.android.cardform.base.CoroutineContextProvider
 import com.mercadolibre.android.cardform.base.Response.Failure
 import com.mercadolibre.android.cardform.base.Response.Success
 import com.mercadolibre.android.cardform.base.resolveRetrofitResponse
@@ -12,8 +12,9 @@ import kotlinx.coroutines.withContext
 
 internal class FinishInscriptionRepositoryImpl(
     private val accessToken: String,
-    private val finishInscriptionService: FinishInscriptionService
-) : BaseCoroutine(), FinishInscriptionRepository {
+    private val finishInscriptionService: FinishInscriptionService,
+    private val contextProvider: CoroutineContextProvider = CoroutineContextProvider()
+) : FinishInscriptionRepository {
 
     override suspend fun getFinishInscriptionData(token: String) =
         withContext(contextProvider.IO) {
