+++ cardform/src/main/java/com/mercadolibre/android/cardform/data/repository/TokenizeRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.mercadolibre.android.cardform.data.repository
 
-import com.mercadolibre.android.cardform.base.BaseCoroutine
+import com.mercadolibre.android.cardform.base.CoroutineContextProvider
 import com.mercadolibre.android.cardform.base.Response.Success
 import com.mercadolibre.android.cardform.base.Response.Failure
 import com.mercadolibre.android.cardform.base.resolveRetrofitResponse
@@ -14,8 +14,9 @@ import kotlinx.coroutines.withContext
 internal class TokenizeRepositoryImpl(
     private val accessToken: String,
     private val tokenizeService: TokenizeService,
-    private val webCardTokenBodyMapper: WebCardTokenBodyMapper
-) : BaseCoroutine(), TokenizeRepository {
+    private val webCardTokenBodyMapper: WebCardTokenBodyMapper,
+    private val contextProvider: CoroutineContextProvider = CoroutineContextProvider()
+) : TokenizeRepository {
 
     override suspend fun tokenizeCard(cardInfoBody: CardInfoBody) =
         withContext(contextProvider.IO) {
