+++ cardform/src/main/res/layout/cf_web_view_progress_state.xml
@@ -15,6 +15,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
+        app:andesProgressTint="@color/ui_components_android_color_accent"
         app:andesProgressStart="true"
         app:andesProgressSize="xlarge"/>
